Include positive feedback when compilation succeeds
Sometimes students were confused by the fact that a successful
compilation yielded no positive message, or by the fact that the
compiler spat out multiple warnings which the students were unable to
distinguish from errors. Therefore, we want Kelvin to return some form
of a positive feedback for the peace of mind of our students.